[TASK] Small performance update for leadlist
As far as every page visits leads to an updated entry in tx_lux_domain_model_lux.tstamp we can also use this field for using the date and time filter. This isn't as much expensive as building joins to pagevisits table.
This change effects also if there is no filter set because the dto always returns values for start and stop