[BUGFIX] Remove non existing field in table definition
Because linkclick table is needed with the next major release (Link Listener Feature) and this is an unused table at the moment, we could also delete the complete table for now. But because this table will be needed in lux 9 it could stay for the moment.

Related: https://github.com/in2code-de/lux/issues/7