[FEATURE] Performance update for lead list view
It turns out that server with low memory_limit and with visitors with a lot of page visits could cause problems in lead list module. The lazy loading of visitor.pagevisits did not work here because this 3 places are forced a relation to pagevisits. We introduced viewhelper with sql queries (over repository) here for a better performance and now lazy loading works.