[TASK] Code cleanup - remove outdated JavaScript function
In the past third party extensions like luxenterprise could use a function in lux to register own functions that can be called as soon as the fingerprint is ready calculated. But that turned out to be a bit dizzy. Imagine a scenario where a fingerprint (or better a value for local storage) is calculated very fast. But maybe the third party JS is not loaded at this point. So nu function could be registered and executed.
We now recommend to look if a fingerprint value is ready calculated in your third party JS and then execute your function by yourself now.