[TASK] Find also leads without any pagevisit relations
In some rare cases visitors are tracked without pagevisits records. That can happen if someone who is opted out (or is using the DNT header) downloads an asset via email4link functionality. Now the visitor is added without pagevisits.
If we now search for email or a part of the email in the lead list, no lead is found. This happened because there is a default from and to date setting (even if no time was chosen in filter).

(cherry picked from commit d27488a2c68232ca2d2af7d6a277f58bc62072b6)