[BUGFIX] Harden lead list and filter posibilities in module
If we choose a time from and a time to, we want to find visitors with pagevisits within this frame and we want to find visitors completely without pagevisits but with a creation date within this frame.
But if we search for a specific searchterm (without further filter settings), we do not want to search only in a frame.