[BUGFIX] Prevent SQL exception in page visit module on some cases
E.g. if you only select a categoryscoring, we don't want to run into an exception. The solution is simple: We have to always create a join to visitor table because this is needed for tables pv and cs - and of coure for v. So let's enable it in general.