[FEATURE] Remove email4link markup from DOM on every page and load it via AJAX
To keep the DOM as small as possible on every page, we now want to get the email4link lightbox content via AJAX. This is done in this commit. Some TypoScript and Template settings were changed with this feature.
We also updated basiclightbox to version 5.

Related: https://projekte.in2code.de/issues/55933