[TASK] Prevent "Type of AbstractModel::$_languageUid must be ?int" error in TYPO3 12
Because TYPO3 12.3 seems to add a typehint to property $_languageUid now, we cannot overrule this property for TYPO3 11 and 12 at the same time (because in TYPO3 11 there is no typehint yet). We now do a workarround with a _setProperty() call in constructor class.
All this to have records with sys_language_uid=-1 in database