[BUGFIX] Fix link to info uri
For whatever reason there were two commits before that should fix this issue and I'm pretty sure I checked the functionality, there was still an issue. I finally (hope so) fixed this issue now. Tested in TYPO3 11 and 12.
In addition the current solution is the cleanest because the already constructed uriBuilder object for the action links can also be used to create another link to another action.