[BUGFIX] Prevent error "Unable to generate a URL for the named route"
when a module should be hidden by configuration. It turned out, that there are links from one LUX module to another LUX module. If the configuration of a module is disabled completely, there cannot be a link to this module if configuration is missing.
With this change we're going to load the configuration of all modules at every time but respect the showing of the module icons via UserTSConfig now.