[BUGFIX] Fix LinkListener record handling for editors
Because basically editors cannot work with records in PID 0, you can define where LinkHandler records can be saved even before this commit in extension manager settings in "pidLinkClickRedords".
But if this page is not of type folder, an error will occur if anyone adds a new LinkListener via Backend Module (1). In addition, it's not possible to select those LinkListeners in the RTE link wizard, because all records from PID 0 will be listed by default (when the page tree is disabled).
In general, I would tend to not show the pagetree in link wizard because I guess that LL records are only stored in one single page. So my first approach was to think about using the option "storagePid" in LinkHandler configuration and pass the value from Extension Manager, but this is not possible (maybe if we would store the value in site configuration instead of the extension manager settings, but this would be a breaking change...).
After that I was wondering if extending PHP class TYPO3\CMS\Recordlist\LinkHandler\RecordLinkHandler with an enforced value "storagePid" would be easy, but this will not work because RecordLinkHandler is a final PHP class and so it cannot be extended. And I don't want to copy the complete class to my own class just for overwriting one single value.
Now I came back to simple render the page tree, to get this issue solved.
From a usability point of view this isn't the best way, but now the records can be used from editors and admins - no matter if they are stored in PID 0 or >0.

So, this commit basically does two things:
1) Allow that table records of type tx_lux_domain_model_linklistener can be saved on normal pages now
2) Disable option "hidePageTree" to show the page tree in link wizard. So now an editor can select the page where LinkListeners are stored and select one

Note: There is still one thing open - The access right cannot be set correctly related to this issue on forge: https://forge.typo3.org/issues/101633

Related: https://projekte.in2code.de/issues/58620