[BUGFIX] Don't enforce join from visitor to pagevisit table if not needed
While I faced an issue with different filter startdates in url shortener list view when searching for a word vs. empty filter, I stumbled over the old workarround 12790994019696a5e654712ecaf3c2c2c1493c68. To fix my problem in the URL shortener view (and for cleaning up reasons), I removed the workarround in FilterDto::getStartTimeFromTimePeriod().
But I wanted to also fix the problem from the old workarround (look at revision above). So, searching for a single searchterm in lead list (e.g. email) should also show visitor records without pagevisits in relation. That's why we fixed it now for admins only, where we do not need a where clause for pagevisits.site if there is not site set in filter. But for editors we need this clause in every case to ensure, that editors can only see authenticated leads. The where-clause leads to a join to pagevisits table and so no leads can be found without pagevisit records.

In addition I also cleaned up a bit the VisitorRepository by using isser functions.