[BUGFIX] ->isFilterSet() should not return true when filter is build with a timePeriod in constructor
Because we stumbled of shortener list view where "clear filter" button was always shown even if there was no filter given. Because before this commit, every timePeriod different to default triggers the isSet() (because isFilterSet() was true), no matter if the value was given via POST param (filter) or from constructor. Now we make a difference between constructor and setter value.
For whatever reason extbase property mapper uses constructor injection in FilterDto if the variable has the same name as a property. To prevent this, because we need to use the setter setTimePeriodDefault(), we need a different variable name for the first parameter in the constructor now.