!!![FEATURE] Define where to save linklistener, shortener and utm-generator records via User TSConfig now
Before this commit, there was a general setting in ext_conf_template.txt file for linklistenr records. And this accidentally was also used for shortener and utm-generator records in LUXenterprise. Now with the introduction of multiclient support, we want to store those records on different pages - depending on the backend user that is working with those tables.
Summary: While the setting in ext_conf_template.txt file was removed, now a User TSConfig setting should be added to be_groups (or be_user) when those records should not be stored on root page (where editors normaly don't have access) like:
tx_lux {
   defaultPage {
      tx_lux_domain_model_linklistener = 1
      tx_luxenterprise_domain_model_shortener = 2
      tx_luxenterprise_domain_model_utmgenerator_uri = 3
   }
}