[TASK] Rename virtual page requests to event tracking request
Because this is more common in marketing. And we don't want to only change the labels, we also want to change all related code stuff to this for better managing in future.
This is basically a breaking change. Nevertheless, we believe that this feature is so new, that nobody is using it at the moment.