[BUGFIX] Prevent sql exception when filtering by category in lead list in MySQL
As long as we use extbase for the query for findAllWithIdentifiedFirst(), there is a problem when trying to order by `categoryscorings.scoring DESC` because MySQL (5 and 8) expects this value also in the select field list. But we cannot manipulate the select fields in extbase repository functions. That's quite funny, because MariaDB 10 works fine here.

It's a pitty, but a general orderings by categoryscoring descending isn't possible in the short run any more here. Maybe we should rewrite the function with pure SQL in the future.

Related: https://github.com/in2code-de/lux/issues/60