[BUGFIX] Hardening of log() functionality
Before this commit an already existing visitor that opens a shortener twice was not redirected anymore. Digging into it results in a 500 without any error message. Problem was caused by $visitor->addLog($log) in log(). We have now rewritten the code to querybuilder methods.
Even if I'm not totally sure what caused the issue in TYPO3 13, I guess querybuilder should be a bit faster then before.